Only set a fake git user if a real one is not set
In order to apply a patch in git, the "git.name" and "git.email"
configuration options are necessary. Previously we would set these
to fake values in the local git repository that we want to patch.
That works well but requies history rewriting when we want to
contribute our patches upstream. The new approach implemented by
this commit is to only set the fake git user name and email if
these values aren't set already (most likely globally, although
locally would do as well).